Mushroom Manager
Added script to hold a reference to all mushroom prefabs we might make, so we can just grab the type of mushroom we need to create from inside of it instead of needing farm manager to have a reference to them all

Updated relevant methods to account for the change to Tile and GameObject, then figuring out if something is a tile afterwards.

Tilling, Planting, watering, harvesting, growing, spreading, and dying all appear to work just fine.