Playtest 2/18/22 Bug fixes
Days are no longer tied to timer, and instead tied to stamina.

Can no longer plant over plants

Crop death is still in Mushroom, but TimeManager properly deletes them and sets their tiles to the default ground sprite so it's obvious the ground needs to be retilled

The three main farming tools and mushrooms cannot be used outside of the farm field, and cannot be used when they wouldn't have an effect, preventing unnecessary stamina loss

Fixed issue with negative item stacks. Now just combine existing stack with a stack of -1, which is just subtraction. When there are no more, the stack is deleted

Changed Dialogue to only use Spacebar