Added Bed; Interaction Update
Bed exists in a rudimentary fashion

Has a sprite and everything. Can be interacted with and you can sleep.

Screen darkens a bit at night

Changed E interaction to Space

Added boolean to check if the player is talking to someone, and uses it as a check to make sure you can't take action when in a conversation.