Tile Saving WIP
Tile Saving is a work in progress. Can, for the most part, go back and reset the field to a previous state after saving it

Ground tiles are all becoming plain, even if they were tilled. And a tile that was only tilled at time of save is apparently moist, which happened after the save.

Red, purple, and glowy were all planted, and apparently saved properly, but only one of them is set when loading

Will not merge to main until properly working