Bug Fixing
Changing things in one scene for an object that exists in multiple is very annoying, since all references have to be fixed. Doing that now, but Fungus is giving me some sort of invocation error in the tutorial flowchart, even though there's nothing with no object.

There was some odd error with a command that I couldn't delete, so it might be related to that. The block was duplicated and the command was deleted, along with the original block.

Restarting laptop to see if that fixes anything, so while this is a WIP, it's being committed.