Basics mostly complete?
The most crucial things that are always present at all times: player gold, the date, and their stamina, correctly go between scenes, it looks like.

Inventory is still weird, but that should be getting worked on.

Most situational things are next like the field, which is only relevant when the player goes to or from their farm. Probably just a check for their farm field to see if it's null. If it is - because there's none in the scene - then do nothing.