NPC, Pet, and Entity Persistence
NPCs appear to persist between scenes now. If you talk to them in a different scene, they say their second piece of dialogue if you already spoke to them before. They appear to remember your quest progress if you go between scenes.

Same with pets; they follow the player from scene to scene

Dropped items remain on the ground when you leave and return, and if you don't pick them up, they're still there for you when you come back

Added type to mushroom, tool, and pet items, so that persistence, as well as saving, can use that information to seek for the proper folder to load things in from. It'll go to the tools folder to load a hoe prefab, but mushrooms to load a mushroom prefab.