Save change; minor changes
If the player saves where there is no farm field, it just saves the persistent scene's version of it, which was the last time it was in a scene, and therefore saved.

That way, they can save on a screen without the farm, then their crops are all still there when they return.

Added Fungus Block to confirm saving, since there was no confirmation of any sort before.

Disabled TutorialObjective text box outside of the tutorial

Saves the scene the player is on when they save, so it puts them back there.