Save Overhaul
Load scene was apparently happening too quickly and erasing instantiated objects that weren't already in the scene.

Made something very similar to Scene Persistence that writes to a file and uses that for loading, setting things to be loaded, loading the scene the player was on first, and then loading things in, instead of basically doing them at the same time.