Gamepad Controls WIP
Stuck on Indicator movement.

Doesn't move unless moving stick, and snaps back when you let go rather than moving it, leaving it there, and leaving it there while you move.

Mouse also only works, as long as you're moving the mouse, since that's the only time the method runs.

Some way is needed to make the right stick (and mouse) move the indicator and leave it there, rather than snapback or have it not move altogether.