NPC Fix, Buff Update, and NPC Template Update
When going into the house to go to bed, the NPC correctly set its day equal to what it was when you last saw them, which was the day before. If the dates are not the same, TimeManager now makes sure they're the same, so NPCs have the proper day at all times.

Regen Buff tested. Just need stats script to replace test int with the player's health.

NPC Template now has showing the hotbar and starting player movement as blocks in Start. All 120 days of dialogue, and all other times when these need to be referenced just call these blocks. There has to be a single reference to the player/hotbar for any errors to be fixed.