Dungeon Prep
Created new scene for dungeons

Populated scene and updated scripts to automatically grab whatever caused errors because of broken references. Should streamline the process of making new scenes in the future

Made Indicator and Main Camera Prefabs, so they could easily be dragged in where needed

Made script so that gold constantly updates, rather than only when the shipping bin is on screen to pay them.