Player Death
When the player dies, they're transported back to their bed and lose whatever items they had obtained in the scene in which they died.

Remove Items didn't work very well if how many you had was exactly double of how many you wanted to get rid of, deleting the slot when it shouldn't. That has been tweaked to prevent that.