Combat Update
Added part of player's strength onto bow damage, so it does more damage the higher the player's level.

Changed sword to do the same, since it was using the enemy's strength to calculate damage, rather than the player's

There was a broken room in Dungeon 1. It has now been fixed.